changed from MD5 to SHA256 (#12)
