+++ AjaxPro/Handler/ConverterJavaScriptHandler.cs
@@ -93,7 +93,7 @@ namespace AjaxPro
 				}
 			}
 
-			etag = MD5Helper.GetHash(System.Text.Encoding.Default.GetBytes("converter"));
+			etag = Hash5Helper.GetHash(System.Text.Encoding.Default.GetBytes("converter"));
 
 			DateTime now = DateTime.Now;
 			DateTime lastMod = new DateTime(now.Year, now.Month, now.Day, now.Hour, now.Minute, now.Second); //.ToUniversalTime();
